Fix display of wrong help message
With the introduction of the --reverse-hidden-helper option that is
inserted into the provided arguments we changed the behavior of the
program when no argument is given. It will complain about a problem with
this hidden option rather than the fact that arguments are missing which
is clearly wrong.
This change inserts the hidden option to reverse the arguments on the
fly only in case the --reverse option is actually supplied.