Factor out replaceFileString function
A different change required the means to replace the string '{file}'
with an actual file name in a given command along with its arguments
from outside of the repository.py file.
Although the previosuly mentioned change got discarded, this change
factors out the replaceFileString function to make it potentially
accessible from the outside and adds a test case for it. It is moved
into a new file, commands.py, because of the expectation of more command
related functionality being added in the foreseeable future.