Factor out checkFileString function
In case the --snapshot-ext option is passed in by the user, we perform a
set of sanity checks. One of those is checks verifies that the a certain
string, '{file}', is contained in the last receive filter (in the case
of a backup operation) or in the first send filter (in case of a
restore).
With the efforts for introducing "spring" support throughout the
program, this change factors out a checkFileString function that
performs this check. Verification is done differently based on whether
the filter commands represent a pipeline or a spring.