Factor out replaceFileString functions
Another task to be performed in a different fashion depending on whether
a list of filter commands is a "simple" pipeline or a spring is the
replacement of '{file}' string with the actual file name when performing
the backup or restore operation.
This change factors out a new function that performs the proper
replacement, depending on the type of the commands to execute. The
behavior is largely different between the two types: in case of a
pipeline we replicate the '{file}' string for all the files we want to
expand it to and perform the replacement. Associated (long) options are
replicated as well. For springs, however, we clone the entire command
and substitute the string. The reason for this behavior is that the
introduction of spring support was based on the premise that certain
programs cannot handle multiple files. So, in such a case we have to
invoke the program multiple times instead--with each instance operating
on a different file.