Introduce runCommands function
In order to fully support springs throughout the program, a different
execution function needs to be invoked depending on whether we are
dealing with a "simple" pipeline or a spring.
This change introduces a new function that, based on the form of a
command, decides whether a pipeline like execution is in order or
whether a spring is desired.