Add end-to-end GnuPG test case
Recent changes made it possible to backup subvolumes in an encrypted
fashion and restoring them later. Encryption and decryption is handled
by the backup program which means that the destination has no insight in
to data being backed up.
One possible program typically used for encryption purposes is GnuPG.
This change adds a new test case that uses this program to backup data
in an encrypted form and restoring it later on.