Fix error when purging enumerated snapshots
When invoking a purge operation on a repository that contains enumerated
snapshots with the same time stamp, the parsing of the time stamp failed
with a ValueException: ValueError: unconverted data remains: -1. The
problem was that the enumeration number is passed to datetime's strptime
method which subsequently cannot parse the entire string and raises an
exception.
This change fixes the problem by pre-parsing the time stamp from the
snapshot name.