Do not parse list of changed files
Until now we parsed the byte stream returned by execution of the btrfs
subvolume find-new command. As it turns out, this is not safe. Cases
were seen where a UnicodeDecodeError was raised because the bytes object
contained characters that could not be decoded properly (such as 0xbb
which produces a combined version of two greater signs ">>"). It is
unclear as to why exactly these bytes cause a decoding error in
conjunction with utf-8.
This change replaces the Repository's diff() method with changed(). The
latter simply reports whether (as opposed to which) files in a subvolume
changed with respect to a snapshot. This way we can directly work on the
bytes object we retrieved as the output of the respective btrfs command
and do not have to decode it. Such a change is possible because nobody
relied on the actually changed files but only the fact whether or not
files change.

Note: I confirmed that the previously used encoding ('utf-8') equaled
      the encoding reported by sys.getfilesystemencoding(). It is not
      entirely clear why files can exist in a different encoding or at
      least containing characters not represented this way in the given
      one. A possible explanation is that these files were copied from a
      file system using a different encoding. However, it does not
      really matter why such cases are possible for purposes of this
      program, the mere fact that there exists one case where it broke
      is enough to justify such a change.