Add remote file repository test case
This change adds a test case to execute the code paths taken when
backing up local data to a remote file repository. The connection to the
remote host is established using SSH. Because we cannot always assume
that the user has an SSH host available this test is dependent on the
setting of the TEST_SSH_HOST (containing the host name as understood by
ssh(1)) environment variable and, thus, optional.
The test itself creates a temporary directory on the remote host that
acts as the file repository's root. It then performs a full backup and
restore cycle using this repository.