Add path encode and decode functionality
One open goal at this stage is a slight change in the implementation: we
want snapshots to be treated as first class objects instead of them
being simple strings (representing their paths).
One reason for of such a change is that the current approach of only
doing a one way conversion (snapshot path to file name) and inferring a
match by mere search for a common prefix can lead to wrong associations
between snapshot names and their actual paths.
As a first step towards achieving this aim, this change introduces the
means to encode and decode (snapshot) paths. This functionality can be
used to create a file name based on a snapshot's path and to retrieve
the path back later on. In a file name, the character '@' (without
quotes) is used for representing a directory separator. If a path
contains this character, it is escaped accordingly in order to always be
able to unambiguously infer back the path.