Fix bug causing invalid parents to be used
Issues have come up where backup failed with different error messages
such as:
  > ERROR: rename o258-8-0 -> bin failed. Directory not empty

Investigation has shown that these are the result of unrelated existing
snapshots of different subvolumes being used as clone sources.
The scenario when such unrelated snapshots are taken into consideration
is when the path of the subvolume to backup is a prefix of existing
other subvolumes of which snapshots are located in the same repository.
In particular, when backing up a btrfs subvolume that is located in
the root ('/') of the file system hierarchy, all other existing
snapshots were used as parents.
This change fixes the problem by taking a trailing separator between the
path encoded in a snapshot's name and its time into consideration when
comparing snapshot names. This trailing separator, which cannot be
contained in a snapshot's name, acts as a suffix making comparisons
based on the base name safe.