Rework automatic argument reversal with --reverse
The --reverse option, if specified, causes other arguments (the source
and destination repository as well as the send and receive filters) to
be swapped. In order to achieve this goal, it relies on a hidden helper
option that is evaluated last and that acts upon the 'reverse' attribute
(set when --reverse is encountered).
Unfortunately, changes made to the argparse module with Python 3.4
modified the implementation in an incompatible way: the namespace
containing the 'reverse' attribute is no longer passed to subparsers
(which we use extensively). Rather, the returned namespace is merged
into the final one (i.e., the one of the "main" parser).
To cope with this incompatibility this patch reworks the handling of the
--reverse option from using a hidden helper option to specifying a
custom "action". This action now performs the argument reversal. The
passing of the namespace of interest to actions was left unchanged
between the Python versions 3.3 and 3.4.