Use action for checking --snapshot-ext constraints
The similarly to the --reverse option, --snapshot-ext used a namespace
that is assumed to be populated with all previously parsed arguments.
With Python 3.4 the behavior of the argparse module was changed and this
constraint is no longer fulfilled.
This change adjusts the --snapshot-ext parameter to use an action for
checking the respective constraints.