Specify parameters for argument reordering
The option handling in the argv module is not correct in the face of an
option without an argument. The problem is that we cannot safely deduce
whether or not an option has an associated argument by solely looking at
the structure of the arguments. In particular, an option that is located
just before the positional arguments will be treated as having an
associated argument, which can result in wrong behavior.
This change fixes the module and with it all invocations by explicitly
requiring the user to state whether or not the option contains an
additional argument or not.
As a bonus, the change also fixes a potential out of index access when
an option to remove (or to reorder) is located at the end of an argument
vector.