Remove --reverse parameter
The --reverse parameter, used for transforming a "backup" operation into
a "restore" by simply adding a parameter was a good idea but makes the
code unnecessarily complex. Also, it is hard to test sufficiently as the
number of parameter combinations that would require testing explodes
with more parameters being added.
This change removes the --reverse parameter and the associated code in
an attempt to simplify the internals slightly.