Handle subvolume mounted under different name correctly
We have seen an interesting case where btrfs-backup would perform a slow
full backup instead of an incremental one although there were existing
backups that would encourage creation of an incremental one.
Investigation has shown that the reason for this behavior is that if a
repository object is created in a subvolume that is mounted under a name
other than the actual subvolume name, the snapshot listing does not work
properly. In particular, the 'btrfs subvolume list' command uses the
internal subvolume name, not the name of the mount point for reporting
snapshots.
This change works around this problem by querying the actual subvolume
name using the 'btrfs subvolume show' command. The name is then used to
craft the anticipated path used in the listing.