Move directory check from _isRoot into _findRoot
The check being performed in the _IsRoot function to see whether the
directory given is an actually existing one does not have to be
performed for every directory in traversed using _findRoot. We only
traverse the list upward (that is, from some directory towards the root
of the file system) and once we convinced ourselve that the initial
directory is valid all subsequent ones should be so as well.
To that end, this change moves the directory check into the _findRoot
function in order to get rid of some unnecessary work but also in
anticipation of a future change which will repurpose _findRoot.