Rework sub-command dispatching
The dispatching of sub-commands has always been a bit messy, with string
comparisons that are prone to become outdated once changes are made. To
that end, this change reworks this dispatching and gets rid of this
problem. We use the subparser's set_defaults method to unconditionally
set the 'method' attribute which can be invoked uniformly by callers.