Make adjstments for btrfs-progs-4.4
With btrfs-progs 4.4.1 the format of two messages was changed, causing
us to fail to find the a btrfs root directory as well as a test failure
due to a string mismatch.
This change addresses both problems. Note that we do not keep backwards
compatibility for we do not have paths accepting the previous strings,
meaning that btrfs-progs version 4.4 (or higher) is a requirement.