Adhere to new root find behavior
Between versions 4.6.1 and 4.10.2 the output emitted in case 'btrfs
subvolume show' is executed on the root volume changed again. We no
longer see a special string but the normal subvolume output. An
additional complication is that sometimes this output contains the
special name '<FS_TREE>' and sometimes it does not. All that causes our
find-root-directory algorithm to no longer work.
This change adjusts the program to correctly handle this new behavior.
It does so by using the 'btrfs filesystem show' command instead of the
subvolume related counterpart. Note that we also drop compatibility with
btrfs-progs versions earlier than 3.17.3 which exhibited yet another
output.