Fix handling of btrfs volumes mounted with 'subvol' option
Until now btrfs-backup was unable to handle cases where the source or
destination volume to work with was mounted with the 'subvol' option
changing the effective root of the volume. The problem with such
scenarios was purely based on the weird listing behavior exhibited by
the 'btrfs subvolume list' command which would always include a
subvolume's path relative to the btrfs root, no matter if this path is
not visible at all in the current file hierarchy.
With this change we have found a way to work around this problem. By
using two commands from the 'btrfs inspect-internal' leaf we are able to
inquire this path.