Fix exception when synchronizing remotely
We are lacking the proper prefixing of the command to issue with the
remote command when synchronizing remotely, potentially resulting in an
exception being raised because of a None being used.
This change fixes the problem by using Repository.command to correctly
prefix the commands issues by the _findSubvolPath function.